requirements.txt
tensorflow==2.9.2
keras==2.9.0
numpy==1.21.6
pandas==1.3.5
scikit-learn==0.23.1
pillow==7.2.0